requirements.txt
-i https://pypi.org/simple
certifi==2018.11.29
chardet==3.0.4
django==2.2.15
django-redis==4.11.0
django-shortuuidfield==0.1.3
furl==2.1.0
idna==2.8
orderedmultidict==1.0.1
pillow==9.2.0
pymysql==0.10.0
python-dateutil==2.7.5
pytz==2021.1
redis==2.10.6
reportlab==3.6.11
requests==2.24.0
shortuuid==1.0.1
six==1.12.0
sqlparse==0.4.2; python_version >= '3.5'
timeconvert==1.5.1
tzlocal==2.1
urllib3==1.24.1
xlrd==1.2.0
